Changed all variables dealing with costs/pocket credit to a long long.
This reflects the total value a pocket can have, which is a postgresql bigint, 8 bytes.

+++ netvend/commands.h
@@ -60,14 +60,14 @@ namespace commands {
     class PocketTransfer : public Command {
         unsigned long fromPocketID_;
         unsigned long toPocketID_;
-        unsigned long amount_;
+        unsigned long long amount_;
     public:
-        PocketTransfer(unsigned long fromPocketID, unsigned long toPocketID, unsigned long amount);
+        PocketTransfer(unsigned long fromPocketID, unsigned long toPocketID, unsigned long long amount);
         void writeToVch(std::vector<unsigned char>* vch);
         static commands::PocketTransfer* consumeFromBuf(unsigned char **ptrPtr);
         unsigned long fromPocketID();
         unsigned long toPocketID();
-        unsigned long amount();
+        unsigned long long amount();
     };
     
     class CreateFile : public Command {
@@ -110,75 +110,75 @@ namespace results {
 
     class Result {
         unsigned char error_;
-        unsigned long cost_;
+        unsigned long long cost_;
     public:
-        Result(unsigned char error, unsigned long cost);
+        Result(unsigned char error, unsigned long long cost);
         virtual void writeToVch(std::vector<unsigned char>* vch);
         static Result* consumeFromBuf(unsigned char **ptrPtr, unsigned char commandType);
-        unsigned long cost();
+        unsigned long long cost();
         bool error();
     };
     
     class Batch {
         commands::Batch* initiatingCommandBatch_;
         std::vector<boost::shared_ptr<Result> > results_;
-        unsigned long cost_;
+        unsigned long long cost_;
     public:
         Batch(commands::Batch* initiatingCommandBatch);
         void addResult(boost::shared_ptr<Result> result);
         void writeToVch(std::vector<unsigned char>* vch);
         void consumeFromBuf(unsigned char **ptrPtr);
         std::vector<boost::shared_ptr<Result> >* results();
-        unsigned long cost();
+        unsigned long long cost();
     };
 
     class CreatePocket : public Result {
         unsigned long pocketID_;
     public:
-        CreatePocket(unsigned long cost, unsigned long pocketID);
+        CreatePocket(unsigned long long cost, unsigned long pocketID);
         void writeToVch(std::vector<unsigned char>* vch);
-        static results::CreatePocket* consumeFromBuf(unsigned long cost, unsigned char **ptrPtr);
+        static results::CreatePocket* consumeFromBuf(unsigned long long cost, unsigned char **ptrPtr);
         unsigned long pocketID();
     };
     
     class RequestPocketDepositAddress : public Result {
         std::string depositAddress_;
     public:
-        RequestPocketDepositAddress(unsigned long cost, std::string depositAddress);
+        RequestPocketDepositAddress(unsigned long long cost, std::string depositAddress);
         void writeToVch(std::vector<unsigned char>* vch);
-        static results::RequestPocketDepositAddress* consumeFromBuf(unsigned long cost, unsigned char **ptrPtr);
+        static results::RequestPocketDepositAddress* consumeFromBuf(unsigned long long cost, unsigned char **ptrPtr);
         std::string depositAddress();
     };
     
     class PocketTransfer : public Result {
     public:
-        PocketTransfer(unsigned long cost);
+        PocketTransfer(unsigned long long cost);
         void writeToVch(std::vector<unsigned char>* vch);
-        static results::PocketTransfer* consumeFromBuf(unsigned long cost, unsigned char **ptrPtr);
+        static results::PocketTransfer* consumeFromBuf(unsigned long long cost, unsigned char **ptrPtr);
     };
     
     class CreateFile : public Result {
         unsigned long fileID_;
     public:
-        CreateFile(unsigned long cost, unsigned long fileID);
+        CreateFile(unsigned long long cost, unsigned long fileID);
         void writeToVch(std::vector<unsigned char>* vch);
-        static results::CreateFile* consumeFromBuf(unsigned long cost, unsigned char **ptrPtr);
+        static results::CreateFile* consumeFromBuf(unsigned long long cost, unsigned char **ptrPtr);
         unsigned long fileID();
     };
     
     class UpdateFileByID : public Result {
     public:
-        UpdateFileByID(unsigned long cost);
+        UpdateFileByID(unsigned long long cost);
         void writeToVch(std::vector<unsigned char>* vch);
-        static results::UpdateFileByID* consumeFromBuf(unsigned long cost, unsigned char **ptrPtr);
+        static results::UpdateFileByID* consumeFromBuf(unsigned long long cost, unsigned char **ptrPtr);
     };
     
     class ReadFileByID : public Result {
         std::vector<unsigned char> fileData_;
     public:
-        ReadFileByID(unsigned long cost, std::vector<unsigned char> fileData);
+        ReadFileByID(unsigned long long cost, std::vector<unsigned char> fileData);
         void writeToVch(std::vector<unsigned char>* vch);
-        static results::ReadFileByID* consumeFromBuf(unsigned long cost, unsigned char **ptrPtr);
+        static results::ReadFileByID* consumeFromBuf(unsigned long long cost, unsigned char **ptrPtr);
         std::vector<unsigned char>* fileData();
     };
 
@@ -203,9 +203,9 @@ namespace errors {
     protected:
         std::string what_;
     public:
-        Error(unsigned char errorType, unsigned long cost, bool fatalToBatch);
+        Error(unsigned char errorType, unsigned long long cost, bool fatalToBatch);
         void writeToVch(std::vector<unsigned char>* vch);
-        static Error* consumeFromBuf(unsigned char errorType, unsigned long cost, unsigned char **ptrPtr);
+        static Error* consumeFromBuf(unsigned char errorType, unsigned long long cost, unsigned char **ptrPtr);
         bool fatalToBatch();
         virtual ~Error() throw() {}
         const char* what() const noexcept;
@@ -214,56 +214,56 @@ namespace errors {
     class ServerLogicError : public Error {
         std::string errorString_;
     public:
-        ServerLogicError(std::string errorString, unsigned long cost, bool fatalToBatch);
+        ServerLogicError(std::string errorString, unsigned long long cost, bool fatalToBatch);
         void setWhat();
         void writeToVch(std::vector<unsigned char>* vch);
-        static ServerLogicError* consumeFromBuf(unsigned long cost, bool fatalToBatch, unsigned char **ptrPtr);
+        static ServerLogicError* consumeFromBuf(unsigned long long cost, bool fatalToBatch, unsigned char **ptrPtr);
         std::string errorString();
     };
     
     class InvalidTargetError : public Error {
         std::string target_;
     public:
-        InvalidTargetError(std::string target, unsigned long cost, bool fatalToBatch);
+        InvalidTargetError(std::string target, unsigned long long cost, bool fatalToBatch);
         void setWhat();
         void writeToVch(std::vector<unsigned char>* vch);
-        static InvalidTargetError* consumeFromBuf(unsigned long cost, bool fatalToBatch, unsigned char **ptrPtr);
+        static InvalidTargetError* consumeFromBuf(unsigned long long cost, bool fatalToBatch, unsigned char **ptrPtr);
         std::string target();
     };
     
     class TargetNotOwnedError : public Error {
         std::string target_;
     public:
-        TargetNotOwnedError(std::string target, unsigned long cost, bool fatalToBatch);
+        TargetNotOwnedError(std::string target, unsigned long long cost, bool fatalToBatch);
         void setWhat();
         void writeToVch(std::vector<unsigned char>* vch);
-        static TargetNotOwnedError* consumeFromBuf(unsigned long cost, bool fatalToBatch, unsigned char **ptrPtr);
+        static TargetNotOwnedError* consumeFromBuf(unsigned long long cost, bool fatalToBatch, unsigned char **ptrPtr);
         std::string target();
     };
     
     class CreditInsufficientError : public Error {
-        unsigned long requiredCredit_;
-        unsigned long availableCredit_;
+        unsigned long long requiredCredit_;
+        unsigned long long availableCredit_;
     public:
-        CreditInsufficientError(unsigned long requiredCredit, unsigned long availableCredit, unsigned long cost, bool fatalToBatch);
+        CreditInsufficientError(unsigned long long requiredCredit, unsigned long long availableCredit, unsigned long long cost, bool fatalToBatch);
         void setWhat();
         void writeToVch(std::vector<unsigned char>* vch);
-        static CreditInsufficientError* consumeFromBuf(unsigned long cost, bool fatalToBatch, unsigned char **ptrPtr);
-        unsigned long requiredCredit();
-        unsigned long availableCredit();
-        unsigned long creditMissing();
+        static CreditInsufficientError* consumeFromBuf(unsigned long long cost, bool fatalToBatch, unsigned char **ptrPtr);
+        unsigned long long requiredCredit();
+        unsigned long long availableCredit();
+        unsigned long long creditMissing();
     };
     
     class CreditOverflowError : public Error {
-        unsigned long pocketCredit_;
-        unsigned long addedCredit_;
+        unsigned long long pocketCredit_;
+        unsigned long long addedCredit_;
     public:
-        CreditOverflowError(unsigned long pocketCredit, unsigned long addedCredit, unsigned long cost, bool fatalToBatch);
+        CreditOverflowError(unsigned long long pocketCredit, unsigned long long addedCredit, unsigned long long cost, bool fatalToBatch);
         void setWhat();
         void writeToVch(std::vector<unsigned char>* vch);
-        static CreditOverflowError* consumeFromBuf(unsigned long cost, bool fatalToBatch, unsigned char **ptrPtr);
-        unsigned long pocketCredit();
-        unsigned long addedCredit();
+        static CreditOverflowError* consumeFromBuf(unsigned long long cost, bool fatalToBatch, unsigned char **ptrPtr);
+        unsigned long long pocketCredit();
+        unsigned long long addedCredit();
         unsigned long long totalCredit();
     };
